fix: reset the messaging-opened latch on sign-out
The latch that gates establishment was module state that nothing cleared,
so after a sign-out and a new sign-in in the same tab the provider would
establish a session for the new actor before that actor opened messaging.
The latch now lives beside the active-session registry, which the logout
cleanup already owns, and is reset there; the public entry points keep
their existing import path.

Addresses CodeRabbit review on src/core/matrix/sessionController.ts:80.
